buttons working

+++ src/Point.hpp
@@ -3,6 +3,6 @@
 #include <Arduino.h>
 
 struct Point {
-  uint8_t x;
-  uint8_t y;
+  short x;
+  short y;
 };
